+++ examples/empty-domains/main.tf
@@ -1,5 +1,5 @@
 provider "aws" {
-  region  = "ap-southeast-1"
+  region = "ap-southeast-1"
 }
 
 module "cert" {